+++ src/main/java/com/example/cashcow_api/dtos/milk/MilkSaleTypeDTO.java
@@ -1,5 +1,6 @@
 package com.example.cashcow_api.dtos.milk;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -9,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class MilkSaleTypeDTO {
     
-    private Double amount;
+    private BigDecimal amount;
 
     private LocalDate createdOn;
 
@@ -17,7 +18,7 @@ public class MilkSaleTypeDTO {
 
     private String type;
 
-    public MilkSaleTypeDTO(Double amount, LocalDate createdOn, Double quantity, String type){
+    public MilkSaleTypeDTO(BigDecimal amount, LocalDate createdOn, Double quantity, String type){
         setAmount(amount);
         setCreatedOn(createdOn);
         setQuantity(quantity);
